feat: Add "Sysmon" item to tray menu with version number
The code changes modify the tray menu in the `setup_tray` function of the `mod.rs` file. The changes add a new item to the tray menu called "Sysmon" with the version number of the application. This enhancement improves the user experience by providing easy access to the version information.